Make GATracer API nicer
